update list cubes and organize module
