+++ bdc-portal/src/app/pages/admin/cubes/cube.interface.ts
