+++ bdc-portal/src/app/pages/admin/cubes/describe-cubes/describe-cubes.component.scss
