+++ bdc-portal/src/app/pages/admin/cubes/logs-cubes/logs-cubes.component.html
@@ -1,0 +1,6 @@
+<div class="container-logs_cubes" *ngIf="cube">
+    <h1>{{ cube }}</h1>
+    <hr />
+
+    logs
+</div>
